feat : healthCheck 추가, redis 설정 추가, mysql 서버 추가

See merge request s10-webmobile1-sub2/S10P12A401!17

+++ backend/mokkoji/src/main/java/online/mokkoji/openvidu/controller/OpenviduController.java
@@ -2,11 +2,16 @@ package online.mokkoji.openvidu.controller;
 
 import io.openvidu.java.client.*;
 import jakarta.annotation.PostConstruct;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import online.mokkoji.openvidu.dto.request.SessionReqDto;
-import online.mokkoji.event.service.EventService;
+import online.mokkoji.common.auth.jwt.util.JwtUtil;
 import online.mokkoji.event.repository.EventRepository;
+import online.mokkoji.event.service.EventService;
+import online.mokkoji.openvidu.dto.request.SessionReqDto;
+import online.mokkoji.result.service.ResultService;
+import online.mokkoji.user.domain.User;
+import online.mokkoji.user.service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -23,6 +28,9 @@ public class OpenviduController {
 
     private final EventService eventService;
     private final EventRepository eventRepository;
+    private final JwtUtil jwtUtil;
+    private final UserServiceImpl userServiceImpl;
+    private final ResultService resultService;
 
 
     @Value("${OPENVIDU_URL}")
@@ -40,8 +48,11 @@ public class OpenviduController {
 
     // Session 생성
     @PostMapping("/api/sessions")
-    public ResponseEntity<Map<String, String>> addSession(@RequestBody(required = false) Map<String, Object> params)
+    public ResponseEntity<Map<String, String>> addSession(@RequestBody(required = false) Map<String, Object> params,
+                                                          HttpServletRequest req)
             throws OpenViduJavaClientException, OpenViduHttpException {
+
+        User user = userServiceImpl.getByProviderAndEmail(jwtUtil.getProvider(req), jwtUtil.getEmail(req));
 
         // request body 객체로 직렬화
         SessionProperties properties = SessionProperties.fromJson(params).build();
@@ -49,16 +60,11 @@ public class OpenviduController {
         //세션 생성
         Session session = openvidu.createSession(properties);
 
-
-        // DB에 담을 유저Id를 받기
-        Long userId = eventService.getUserId(params);
-
-
         // DB에 저장할 Dto 생성
-        SessionReqDto sessionReqDto = new SessionReqDto(userId, session.getSessionId(), session.createdAt());
+        SessionReqDto sessionReqDto = new SessionReqDto(user.getId(), session.getSessionId(), session.createdAt());
 
         // DB에 저장
-        String sessionDBId = eventService.createSession(sessionReqDto);
+        eventService.createSession(sessionReqDto);
 
         // 리턴값(sessionId) 담는 map 생성
         Map<String, String> response = new HashMap<>();
@@ -72,9 +78,15 @@ public class OpenviduController {
 
     // Session 삭제
     @DeleteMapping("/api/sessions/{sessionId}")
-    public ResponseEntity<?> deleteSession(@PathVariable("sessionId") String sessionId,
-                                           @RequestBody(required = false) SessionReqDto sessionReqDto)
+    public ResponseEntity<String> deleteSession(@PathVariable("sessionId") String sessionId,
+                                                HttpServletRequest req,
+                                                @RequestBody(required = false) SessionReqDto sessionReqDto)
             throws OpenViduJavaClientException, OpenViduHttpException {
+
+        if (sessionReqDto.getAuthority().equals("sub")) return new ResponseEntity<>("참여자 회의 나감", HttpStatus.OK);
+
+        User user = userServiceImpl.getByProviderAndEmail(jwtUtil.getProvider(req), jwtUtil.getEmail(req));
+        sessionReqDto.setUserId(user.getId());
 
         Session activeSession = openvidu.getActiveSession(sessionId);
         eventService.deleteSession(sessionId, sessionReqDto);
@@ -82,10 +94,11 @@ public class OpenviduController {
         activeSession.close();
 
 
-        // TODO: 2024.01.28 redis에 남은 결과물 파일 저장(resultService)||아직 작성 중인 사람은?
+        // TODO: 2024.01.31 아직 작성 중인 사람은?
+        resultService.saveRemainingPhotos();
+        resultService.saveRemainingMessages();
 
-
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>(" 세션 삭제 완료", HttpStatus.NO_CONTENT);
     }
